MODLOGSAML-107: Delete configuration cache on internal error
Remove the Saml2Client (the configuration) from SamlConfigHolder
(the configuration cache) if there is an internal error that
might be caused by a bogus configuration.

The issue can be reproduced with an IdP metadata XML with
entityID set to an empty string.

To see the actual cause the IdP metadata will be logged on
internal error.

(cherry picked from commit 24e45d43939fe926d7b600221ea5d1d154bac31e)