MODLOGSAML-192: Allow callback endpoint to return RTR tokens when configured (merge to master) (#180)
Currently the callback-with-expiry endpoint is the only way to get the RTR tokens into the client when using SSO in FOLIO. This works, but it requires that the IdP metadata be downloaded from FOLIO and updated in the IdP since the callback is part of this metadata. This adds a new configuration property called useSecureTokens in the SAML configuration to allow, when present and true, for the /callback endpoint to return the RTR more secure expiring tokens in addition to the legacy non-expiring tokens.